Return queue-prefix in open properties
The `queue-prefix` property of the `open` frame is not part of the AMQP spec.
However, let's include this property because it is also returned by ActiveMQ
and Solace and understood by some client libs, e.g. ActiveMQ NMS.AMQP and Qpid JMS.

Note that we do not set `topic-prefix` here because `/exchanges/amq.topic/` is a valid
target address prefix but not a valid source address prefix.

Closes #12531