GitHub Actions: Relax the regex to match the generix-unix archive
[Why]
Our alpha packages published to `rabbitmq/server-packages` use letters
in addition to digits and dots.

[How]
Accept any string that starts with a digit as the version here. We just
enforce the prefix and the filename extension.

The digit at the beginning it here to exclude the
`rabbitmq-server-generic-unix-latest-toolchain-*.tar.xz` archive.