Highlight busy incoming links
Visualise busy links from publisher to RabbitMQ. If the link credit
reaches 0, we set a yellow background colour in the cell.
Note that these credit values can change many times per second while the
management UI refreshes only every few seconds. However, it may still
give a user an idea of what links are currently busy.

We use yellow since that's consistent with the `flow` state in AMQP
0.9.1, which is also set to yellow.

We do not want want to highlight **outgoing** links with credit 0 as
that might be a paused consumer, and therefore not a busy link.

We also use yellow background color if incoming-window is 0 (in case of
a cluster wider memory or disk alarm) or if remote-incoming-window is 0
as consumers should try to keep their incoming-window open and instead
use link credit if they want to pause consumption.

Additionaly we set a grey background colour for the `/management`
address just to highlight them slightly since these are "special" link
pairs.