Switch AWS peer discovery tests from Bazel to make
This change includes some fixes like setting the suffix for the metadata
which is necessary since 6692fcb6. The main reason for this change
though is to switch from running the integration_SUITE from the
AWS peer discovery plugin from Bazel to make, a necessary step to
dropping the Bazel files.