Add back metadata labels to the OCI build
This fixes a regression from 6692fcb6: setting the `labels` to the
output of the metadata step attaches useful standardized labels like
the git hash when built, the source URL and the created timestamp.
Actions like `int128/wait-for-docker-image-action` used in the
`peer-discovery-aws` workflow use the revision label to detect that the
latest image has the desired hash.

Without setting `labels` we seem to inherit the version label from the
base erlang image:

    $ docker image inspect -f '{{ json .Config.Labels }}' \
        pivotalrabbitmq/rabbitmq:sha-0e7b53c6a8b682411c3f0024691a4760d8219699-otp27
    {"org.opencontainers.image.version":"27.2.1"}