Adding a "build-dist" target to the Makefile
This target is identical to the existing "source-dist" target, except that it allows for packaging and testing of the source archive. This is done by including the packaging/ and tests/ directories in the output tarball, along with specific subdirectories that are required by tests.

Signed-off-by: Kartik Ganesh <gkart@amazon.com>
(cherry picked from commit aa9e0a5a280b5d8f4c1e03c4f8d6c522ab509ee3)